requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.2
future==0.18.3
netmiko==4.1.2
ntc-templates==3.3.0
paramiko==3.1.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.0.0
PyYAML==6.0
scp==0.14.5
six==1.16.0
tenacity==8.2.2
textfsm==1.1.2
